refactor(opencode): split Agent Swarm modules
- extract launcher command execution from the Agent Swarm entrypoint

- isolate Agent Builder mode E2E coverage and shared helpers
